agent/requirements.txt
fastapi
uvicorn
openai
python-dotenv
httpx
llama-cpp-python